requirements.txt
# AI Glass System - Python Dependencies
# Python 3.9 - 3.11 supported

# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
starlette==0.27.0

# Computer Vision & Deep Learning
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
ultralytics==8.3.200
torch==2.0.1
torchvision==0.15.2

# MediaPipe (Hand Detection)
mediapipe==0.10.8

# Audio Processing
pyaudio==0.2.14
pydub==0.25.1
pygame==2.5.2

# Aliyun DashScope SDK (ASR & Qwen-Omni)
dashscope==1.14.1
openai==1.3.5  # For DashScope compatibility mode

# Environment & Configuration
python-dotenv==1.0.0

# Utilities
opencv-contrib-python==4.8.1.78  # Extended OpenCV modules

# HTTPX for Ollama async HTTP calls
httpx==0.24.1

# Optional Performance Optimizations
# Uncomment if needed:
# onnxruntime-gpu==1.16.3  # For ONNX model acceleration
# tensorrt==8.6.1  # For TensorRT optimization

# Development & Testing (Optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# flake8==6.1.0

# Platform-specific dependencies
# Windows:
# - PyAudio requires separate installation: https://www.lfd.uci.edu/~gohlke/pythonlibs/#pyaudio
# Linux:
# - Install PyAudio dependencies: sudo apt-get install portaudio19-dev python3-pyaudio
# - Install OpenCV dependencies: sudo apt-get install libgl1-mesa-glx
# macOS:
# - Install PyAudio: brew install portaudio && pip install pyaudio

# CUDA Dependencies (GPU acceleration)
# - CUDA Toolkit 11.8+: https://developer.nvidia.com/cuda-downloads
# - cuDNN 8.6+: https://developer.nvidia.com/cudnn
# PyTorch with CUDA support:
# pip install torch==2.0.1+cu118 torchvision==0.15.2+cu118 --index-url https://download.pytorch.org/whl/cu118

